Caledonia: splitting prehashed algo from lib.rs
